functions/requirements.txt
firebase-admin>=6.2.0
firebase-functions>=0.1.0
flask>=2.0.0
flask-login>=0.6.2
gunicorn>=20.1.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
